requirements.txt
fastapi
uvicorn
google-generativeai
faiss-cpu
sentence-transformers
python-multipart
twilio